repo, archive, restore, etl, workflow

+++ recsys/core/dal/repo.py
@@ -11,7 +11,7 @@
 # URL        : https://github.com/john-james-ai/Recommender-Systems                                #
 # ------------------------------------------------------------------------------------------------ #
 # Created    : Monday November 14th 2022 01:27:04 am                                               #
-# Modified   : Friday November 25th 2022 05:33:50 pm                                               #
+# Modified   : Sunday November 27th 2022 03:29:20 am                                               #
 # ------------------------------------------------------------------------------------------------ #
 # License    : MIT License                                                                         #
 # Copyright  : (c) 2022 John James                                                                 #
@@ -20,9 +20,10 @@
 import logging
 import os
 from abc import ABC, abstractmethod
-from typing import Any
+from typing import Any, Union, List
 import shutil
 
+from recsys.config.data import STAGES
 from recsys.core.dal.dataset import Dataset
 from recsys.core.dal.registry import DatasetRegistry
 from recsys.core.services.io import IOService
@@ -36,32 +37,48 @@ class Repo(ABC):
     """Repository base class"""
 
     @abstractmethod
-    def add(self, *args, **kwargs) -> None:
+    def add(self, *args, **kwargs) -> None:  # pragma: no cover
         pass
 
     @abstractmethod
-    def get(self, id: str) -> Any:
+    def get(self, id: str) -> Any:  # pragma: no cover
         pass
 
     @abstractmethod
-    def remove(self, id: str) -> None:
+    def remove(self, id: str) -> None:  # pragma: no cover
         pass
 
     @abstractmethod
-    def exists(self, id: str) -> bool:
+    def exists(self, id: str) -> bool:  # pragma: no cover
         pass
 
 
 # ------------------------------------------------------------------------------------------------ #
 class DatasetRepo(Repo):
-    """Dataset Repository"""
+    """Dataset Repository
+
+    Repository of Dataset objects.
+
+    Args:
+        directory (str): The base directory for the repository.
+        io (IOService): Service responsible for reading and writing data to, from the repository.
+        file_format (str): The format in which the Dataset objects are stored.
+        archive (bool): Whether to archive rather than delete objects on reset and remove.
+    """
 
     def __init__(
-        self, directory: str, io: IOService, registry: DatasetRegistry, file_format: str
+        self,
+        directory: str,
+        io: IOService,
+        registry: DatasetRegistry,
+        file_format: str,
+        archive: bool = True,
     ) -> None:
         self._directory = directory
         self._io = io
         self._registry = registry
+        self._archive = archive
+        self._archive_directory = os.path.join(self._directory, "archive")
         self._file_format = file_format
 
     def __len__(self) -> int:
@@ -71,14 +88,52 @@ class DatasetRepo(Repo):
     def directory(self) -> str:
         return self._directory
 
-    def reset(self) -> None:
-        msg = "Are you SURE. This will delete EVERYTHING in this repository. If you are sure, type 'YES, I AM SURE': "
-        sure = input(msg)
-        if sure == "YES, I AM SURE":
-            logger.info(f"\n\nRESETTING REPOSITORY at {self._directory}.")
-            self._registry.reset()
-            shutil.rmtree(self._directory, ignore_errors=True)
-            logger.info(f"\n\nREPOSITORY at {self._directory}, RESET.")
+    @property
+    def archive(self) -> bool:
+        return self._archive
+
+    @archive.setter
+    def archive(self, archive: bool) -> None:
+        self._archive = archive
+
+    @property
+    def archive_directory(self) -> str:
+        return self._archive_directory
+
+    def reset(self, purge_archive: bool = False, silent: bool = False) -> None:
+        """Resets the repository and optionally purges the archive.
+
+        Args:
+            purge_archive (bool): Indicates whether the archive should be purged.
+            silent (bool): If True, confirmation prompts are suppressed.
+        """
+        if silent:
+            if purge_archive:
+                self._reset_all()
+            elif self._archive:
+                self._reset_and_archive()
+            else:
+                self._reset_repo()
+        else:
+            if purge_archive:
+                msg = (
+                    "This action will PERMANENTLY DELETE all data and archive. Are you SURE? (Y/N)"
+                )
+                sure = input(msg)
+                if "y" in sure.lower():
+                    self._reset_all()
+            elif self._archive:
+                msg = "Resetting the repository will archive all data. Are you SURE? (Y/N)"
+                sure = input(msg)
+                if "y" in sure.lower():
+                    self._reset_and_archive()
+            else:
+                msg = (
+                    "Resetting the repository will PERMANENTLY DELETE all data. Are you SURE? (Y/N)"
+                )
+                sure = input(msg)
+                if "y" in sure.lower():
+                    self._reset_repo()
 
     def add(self, dataset: Dataset) -> None:
         """Adds a Dataset to the repo.
@@ -103,18 +158,36 @@ class DatasetRepo(Repo):
         """
         try:
             registration = self._registry.get(id)
-            return self._io.read(registration["filepath"])
+            try:
+                return self._io.read(registration["filepath"])
+            except FileNotFoundError as e:
+                msg = f"The registered Dataset {id} data not found in repository. The database may be corrupt.\n{e}"
+                logger.error(msg)
+                raise FileNotFoundError(msg)
         except FileNotFoundError:
             msg = f"No Dataset with id = {id} exists."
             logger.error(msg)
             raise FileNotFoundError(msg)
 
-    def find_dataset(self, name: str, stage: str = None):
+    def get_dataset(self, name: str, stage: str = None):
+        """Returns the latest version of the Dataset that matches the criteria.
+
+        Args:
+            name (str): Name of the Dataset object.
+            stage (str): Preprocessing stage. One of 'staged', 'interim' or 'final'
+        """
+        result = self.find_dataset(name=name, stage=stage)
+        logger.error(f"\n\nget_dataset method returned:\n{result}\n")
+        if result.shape[0] > 1:
+            result = result.sort_values(by="version", ascending=False).reset_index().iloc[0]
+        return self.get(int(result["id"]))
+
+    def find_dataset(self, name: str, stage: str = None, include_archive: bool = True) -> Dataset:
         """Returns a Dataframe of Dataset objects that match the criteria
 
         Args:
             name (str): Required name of Dataset.
-            stage (str): Optional, one of 'raw', 'interim', or 'cooked'.
+            stage (str): Optional, one of 'input', 'interim', or 'final'.
         """
         return self._registry.find_dataset(name=name, stage=stage)
 
@@ -129,9 +202,11 @@ class DatasetRepo(Repo):
         """
         try:
             registration = self._registry.get(id)
-            os.remove(registration["filepath"])
+            if self._archive:
+                self.archive_dataset(registration=registration)
+            else:
+                os.remove(registration["filepath"])
             self._registry.remove(id)
-
         except FileNotFoundError:
             if ignore_errors:
                 pass
@@ -152,9 +227,54 @@ class DatasetRepo(Repo):
         Args:
             dataset (Dataset): The dataset to check.
         """
-        return self._registry.version_exists(
-            name=dataset.name, stage=dataset.stage, version=dataset.version
-        )
+        return self._registry.version_exists(dataset=dataset)
+
+    def archive_dataset(self, id: int) -> None:
+        """Moves a Dataset to archive.
+
+        Args:
+            id (int): Dataset id
+
+        Raises FileNotFoundError if Dataset doesn't exist.
+        """
+        registration = self._registry.get(id)
+        source = registration["filepath"]
+        destination = self._get_archive_filepath(registration)
+        self._move_file(source=source, destination=destination)
+        self._registry.archive(id)
+
+    def restore_dataset(self, id: int) -> None:
+        """Restores a Dataset from archive.
+
+        Args:
+            id (int): Dataset Id
+        """
+        registration = self._registry.get(id)
+        source = self._get_archive_filepath(registration)
+        destination = registration["filepath"]  # Filepath doesn't change under archive.
+        self._move_file(source=source, destination=destination)
+        self._registry.restore(id)
+
+    def archive_repo(self) -> None:
+        """movies repository data to an archive folder."""
+        i = 0
+
+        registrations = self._registry.get_all(fmt=dict)
+        for id, registration in registrations.items():
+            self.archive_dataset(id)
+            i += 1
+
+        logger.info(f"Archive complete, {str(i)} files archived.")
+
+    def restore_repo(self) -> None:
+        """Restores a repository from archive"""
+        i = 0
+
+        registrations = self._registry.get_archive()
+        for id, registration in registrations.items():
+            self.restore_dataset(id)
+            i += 1
+        logger.info(f"Restore complete, {str(i)} files restored.")
 
     def print(self, id: int) -> None:
         """Prints a Dataset by id.
@@ -173,8 +293,9 @@ class DatasetRepo(Repo):
             raise FileNotFoundError(msg)
 
     def print_registry(self) -> None:
-        """Prints list of dataset names in repository."""
+        """Prints and returns a DataFrame containing the registry."""
         print(self._registry.get_all())
+        return self._registry.get_all()
 
     def _get_filepath(self, dataset: Dataset) -> str:
         """Formats a filename using the Dataset name, stage, and version."""
@@ -188,3 +309,54 @@ class DatasetRepo(Repo):
             + self._file_format
         )
         return os.path.join(self._directory, dataset.stage, filename)
+
+    def _get_archive_filepath(self, registration: str) -> str:
+        """Returns the archive filepath given a registration."""
+        return os.path.join(
+            self._archive_directory,
+            registration["stage"],
+            os.path.basename(registration["filepath"]),
+        )
+
+    def _reset_all(self) -> None:
+        """Resets repo and archive. All data is deleted."""
+        self._registry.reset()
+        self._purge_repo()
+        self._purge_archive()
+
+    def _reset_and_archive(self) -> None:
+        """Resets the repository, archiving all data."""
+        self._registry.reset()
+        self._purge_repo()
+        self._purge_archive()
+
+    def _reset_repo(self) -> None:
+        """Resets repo and deletes data. Archive is untouched."""
+        self._registry.reset()
+        self._purge_repo()
+
+    def _purge_repo(self) -> None:
+        """Deletes the data in the repository."""
+        folders = []
+        for stage in STAGES:
+            folders.append(os.path.join(self._directory, stage))
+        self._purge(folder=folders)
+
+    def _purge_archive(self) -> None:
+        self._purge(folder=self._archive_directory)
+
+    def _purge(self, folder: Union[str, List[str]]) -> None:
+        """Purges the designated folder(s) and contents.
+
+        Args:
+            folder (Union[str, List[str]]): Folder or list of folders to purge.
+        """
+        if isinstance(folder, str):
+            shutil.rmtree(folder, ignore_errors=True)
+        else:
+            for f in folder:
+                shutil.rmtree(f, ignore_errors=True)
+
+    def _move_file(self, source, destination) -> None:
+        os.makedirs(os.path.dirname(destination), exist_ok=True)
+        shutil.move(src=source, dst=destination)
